favorite-Ajax

+++ app/controllers/public/favorites_controller.rb
@@ -1,17 +1,22 @@
 class Public::FavoritesController < ApplicationController
   before_action :authenticate_customer!
-  
+
   def create
-    post_video = PostVideo.find(params[:post_video_id])
-    favorite = current_customer.favorites.new(post_video_id: post_video.id)
-    favorite.save
-    redirect_to request.referer
+    @post_video = PostVideo.find(params[:post_video_id])
+    favorite = @post_video.favorites.new(customer_id: current_customer.id)
+    if favorite.save
+    else
+      redirect_to request.referer
+    end
   end
 
   def destroy
-    post_video = PostVideo.find(params[:post_video_id])
-    favorite = current_customer.favorites.find_by(post_video_id: post_video.id)
-    favorite.destroy
-    redirect_to request.referer
+    @post_video = PostVideo.find(params[:post_video_id])
+    favorite = @post_video.favorites.find_by(customer_id: current_customer.id)
+    if favorite.present?
+      favorite.destroy
+    else
+      redirect_to request.referer
+    end
   end
 end
